requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.31
pymysql==1.1.1
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
bcrypt==4.0.1
python-dotenv==1.0.1